change bucket size units to Long

+++ src/main/java/com/auth0/jwk/Bucket.java
@@ -15,9 +15,10 @@ interface Bucket {
     /**
      * Calculates the wait time before the given amount of tokens will be available in the Bucket.
      *
+     * @param count the amount of tokens to check how much time to wait for.
      * @return the wait time in milliseconds in which the given amount of tokens will be available in the Bucket.
      */
-    long willLeakIn(int count);
+    long willLeakIn(long count);
 
     /**
      * Tries to consume one token from the Bucket.
@@ -29,7 +30,8 @@ interface Bucket {
     /**
      * Tries to consume the given amount of tokens from the Bucket.
      *
+     * @param count the amount of tokens to try to consume.
      * @return true if it could consume the given amount of tokens or false if the Bucket doesn't have that amount of tokens available now.
      */
-    boolean consume(int count);
+    boolean consume(long count);
 }
